Move packed encoding functions into separate module
