+++ eth_abi/__init__.py
@@ -4,9 +4,7 @@ from eth_abi.abi import (  # NOQA
     decode_single,
     decode_abi,
     encode_single,
-    encode_single_packed,
     encode_abi,
-    encode_abi_packed,
     is_encodable,
 )
